added userName for language

+++ src/main/java/com/shovov/springsecurity/service/LanguageServiceImpl.java
@@ -4,10 +4,14 @@ import com.shovov.springsecurity.model.Language;
 import com.shovov.springsecurity.model.interfaces.LanguageRepository;
 import com.shovov.springsecurity.service.interfaces.LanguageService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
+import java.security.Principal;
 import java.time.LocalDateTime;
 import java.util.ArrayList;
+import java.util.ConcurrentModificationException;
 import java.util.List;
 import java.util.Optional;
 
@@ -15,14 +19,15 @@ import java.util.Optional;
 public class LanguageServiceImpl implements LanguageService {
     @Autowired LanguageRepository languageRepository;
 
-    private Language saveLanguage(Language language){
-        if(language.getId() != null) language.setLastModified(LocalDateTime.now());
-        return languageRepository.save(language);
+    @Override
+    public String getLoggedInUserName(){
+        return SecurityContextHolder.getContext().getAuthentication().getName();
     }
 
     @Override
     public Language postLanguage(Language language){
-        return saveLanguage(language);
+        language.setUserName(getLoggedInUserName());
+        return languageRepository.save(language);
     }
 
     @Override
@@ -37,25 +42,29 @@ public class LanguageServiceImpl implements LanguageService {
         return languageRepository.findById(id);
     }
 
+    Optional<Language> validateLanguageChangeRequest(long id){
+        Optional<Language> optionalLanguage = languageRepository.findById(id);
+        return (optionalLanguage.isPresent() && optionalLanguage.get().getUserName().equals(getLoggedInUserName()) )
+                ? optionalLanguage : Optional.ofNullable(null);
+    }
     @Override
     public void deleteLanguageById(long id){
-         languageRepository.deleteById(id);
+        Optional<Language> optionalLanguage = validateLanguageChangeRequest(id);
+        if(optionalLanguage.isPresent()) languageRepository.deleteById(id);
     }
 
     @Override
     public Optional<Language> postLanguageById(long id,Language language){
-
-        Optional<Language> optionalLanguage = languageRepository.findById(id);
+        Optional<Language> optionalLanguage = validateLanguageChangeRequest(id);
 
         if(!optionalLanguage.isPresent()){
-            System.out.println("FUca");
            return optionalLanguage;
         }
 
         Language prevLanguageDeltails = optionalLanguage.get();
 
         if(language.getLanguageName() != null) prevLanguageDeltails.setLanguageName(language.getLanguageName());
-        return Optional.ofNullable(saveLanguage(prevLanguageDeltails));
+        return Optional.ofNullable(languageRepository.save(prevLanguageDeltails));
     }
 }
